Bugfix in GFBookkeeping's copy constructor.
Due to the fact that the assignment operator for TMatrixT<double> crashes if
one tries to assign a defined matrix to a uninitialized matrix, it will fail
because the Nrows and Ncols do not match. A resize was added before the
assignment to circumvent this problem. Might become obsolete if the behaviour
of ROOT changes.