Fixed missing/wrong DoxyGen "end of file"-line.
Don't worry: only comments were changed! A quick `grep "@}" $GENFIT/core`
revealed that there seem to be different opinions about how this should be
worded. The two possibilities are "/** @} */" and "/* @} **/". The first one
seems to be correct according to google and to my gut feeling, so I changed
all to the first option.