Finished the getSmoothedHitCoords function in GFTools.
The problem mentioned before was "solved" by just skipping the check in
RKTrackRep.cxx. If I interpret the code correctly, it is not needed anyway.
The rest of the math was added, so now getSmoothedHitCoords should really
do what its name suggests, but I still have to modify the test programm to
produce meaningful histograms.