Rewrote GFTools and SmoothingPulls. Works now!
Rewrote GFTools to allow wider access to the smoothed data. SmoothingPulls now
makes use of this and the error for weighting the pulls is calculated correctly.
Like this, the pulls look fine and smoothing seems to be working. Still to do:
- document the getAuxInfo/hasAuxInfo mechanism
- calculate the smoothed data also for the first and last hit
- introduce biased residuals
In addition, general clean up and documentation should be added.