Genereal clean up.
Removed commented code and unused files. While going back to the initial
smearing of the initialization values for the TrackRep in the smoothing test,
I found that enabling the smearing for the momentum negatively influences the
pulls (sigma > 1) so I left it commented for the moment. Will have to
investigate.