Fixed a bug in GFKalman.
During GFKalman::fittingPass, the failed hits were cleared always for the cardinal track
representation and during every call of GFKalman::fittingPass, even if the track
representation already had status flag 1. Changed the structure a little, now the
clearing of the failed hit is done for the correct track representation and only if the
representation has status flag 1.