Fixed Bug 3241470.
GenfitDisplay now does a GFAbsTrackRep::extrapolate(GFDetPlane&) before getting
the hit position, which leads to the state vector and covariance matrix being
updated in the track representation. For the subsequent hit, the extrapolation
then starts from the previous hit, which means the extrapolation only goes from
hit to hit and not always from the beginnig to hit. In order to leave the track
behind in the same state it was at the beginning, the initial data is stored
and written back once all hits of the track were read out.