Extended the eventDisplay to draw smoothed tracks.
If a track was fitted with smoothing enabled, this information will be used by
the eventDisplay. Also added the possibility to get the auxInfo out of smoothed
tracks with GFTools, which was necessary to get the info into the eventDisplay
in a clean way. I do not like the structure of GFTools very much at the moment,
it has quite some dublicated code because of the 'biased' functionality.