Complete rework again.
Changed the structure of almost everything DAF related. Small change in GFTools,
not worth mentioning and added a releaseTrackReps to GFTrack, to let the Reps go
without deleting them, which I need in the DAF. Introduced some noise hits in
the test and the DAF seems to filter them reliably. Further evaluation will
follow.