Continuing on the DAF and its test.
Simplified GFDaf::processTrack() to make it more readable. The test now uses
several track reps, which seems to work. GFDafHit::clone() still needs some
work, quite some optimization potential in getHitCov and getHitCoord.
GFDaf::processTrack() also needs some rework, as it is not really efficient in
its present state. The test has to be extended and histograms added to show the
performance of the DAF.