First version.
The principle idea is to save the extrapolated states and covs to speed up the
retrieval of smoothed points. To ensure backwards compatibility, it is checked
which keys are saved in the bookkeeping. Currently, all new stuff is commented
out which causes GFTools::getSmoothedData to use the old method. The new method
at the moment gives different results, which is of course bad, also, the
equivalent in the biased method is missing. Probably, the state and cov have to
be saved both before and after the update for this to work, which then
increases memory requirements, so one should be able to choose which method to
use.