Finalized the fast smoothing for unbiased data.
Now one can choose with a setter if one wants to have the faster but more
memory-demanding method or the slower one which requires slightly less space.
This unfortunately required to up the ClassDef version of the GFTrack by one.
Both methods now return the same result. Speed-up should be significant if the
track contains mostly planar hits (plane of foreward and backward fit are the
same).